Move 'document' to a new location under a new name
