+++ public/files/js/charts/lineSelection.ts
@@ -24,7 +24,7 @@
 
 import * as d3 from 'vendor/d3';
 import * as d3Color from 'vendor/d3-color';
-import {PageModel} from '../pages/document';
+import {PageModel} from '../app/main';
 import {MultiDict} from '../util';
 import * as RSVP from 'vendor/rsvp';
