+++ public/files/js/pages/subcorpForm.ts
@@ -25,7 +25,7 @@
 /// <reference path="../vendor.d.ts/react.d.ts" />
 
 import * as RSVP from 'vendor/rsvp';
-import {PageModel} from './document';
+import {PageModel} from '../app/main';
 import {init as subcorpViewsInit} from 'views/subcorp/forms';
 import {SubcorpWithinFormStore, SubcorpFormStore} from '../stores/subcorp/form';
 import liveAttributes from 'plugins/liveAttributes/init';
