+++ public/files/js/pages/view.ts
@@ -28,7 +28,7 @@
 import * as SoundManager from 'vendor/SoundManager';
 import * as RSVP from 'vendor/rsvp';
 
-import {PageModel} from './document';
+import {PageModel} from '../app/main';
 import {MultiDict, parseUrlArgs, updateProps} from '../util';
 import * as conclines from '../conclines';
 import {init as concViewsInit, ConcordanceView} from 'views/concordance/main';
