+++ public/files/js/plugins/defaultAuth/init.ts
@@ -22,7 +22,7 @@
 /// <reference path="../../vendor.d.ts/rsvp.d.ts" />
 
 import {SimplePageStore} from '../../stores/base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import * as RSVP from 'vendor/rsvp';
 
 /**
