+++ public/files/js/stores/coll/collForm.ts
@@ -24,7 +24,7 @@
 import * as Immutable from 'vendor/immutable';
 import * as RSVP from 'vendor/rsvp';
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {MultiDict} from '../../util';
 
 export type AttrValue = {n:string; label:string};
