+++ public/files/js/stores/concordance/lineSelection.ts
@@ -25,7 +25,7 @@
 import {MultiDict} from '../../util';
 import {SimplePageStore} from '../base';
 import {ConcLinesStorage} from '../../conclines';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {ConcLineStore} from './lines';
 import * as RSVP from 'vendor/rsvp';
