+++ public/files/js/stores/concordance/lines.ts
@@ -28,7 +28,7 @@
 
 import {MultiDict} from '../../util';
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import * as Immutable from 'vendor/immutable';
 import {Line, LangSection, KWICSection, TextChunk} from './line';
 import * as RSVP from 'vendor/rsvp';
