+++ public/files/js/stores/freqs/ctFreqForm.ts
@@ -22,7 +22,7 @@
 /// <reference path="../../vendor.d.ts/immutable.d.ts" />
 
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import * as Immutable from 'vendor/immutable';
 import {MultiDict} from '../../util';
