+++ public/files/js/stores/freqs/dataRows.ts
@@ -24,7 +24,7 @@
 /// <reference path="../../vendor.d.ts/rsvp.d.ts" />
 
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import * as Immutable from 'vendor/immutable';
 import * as RSVP from 'vendor/rsvp';
 import {FreqFormInputs} from './freqForms';
