+++ public/files/js/stores/freqs/save.ts
@@ -23,7 +23,7 @@
 
 import * as Immutable from 'vendor/immutable';
 import {SimplePageStore, validateGzNumber, validateNumber} from '../../stores/base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {MultiDict} from '../../util';
 import {ContingencyTableStore} from './ctable';
 import {CTFlatStore} from './flatCtable';
