+++ public/files/js/stores/options/general.ts
@@ -24,7 +24,7 @@
 
 import {SimplePageStore} from '../base';
 import * as Immutable from 'vendor/immutable';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {MultiDict} from '../../util';
