+++ public/files/js/stores/options/structsAttrs.ts
@@ -25,7 +25,7 @@
 
 import {SimplePageStore} from '../base';
 import * as Immutable from 'vendor/immutable';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import * as RSVP from 'vendor/rsvp';
