+++ public/files/js/stores/query/main.ts
@@ -27,7 +27,7 @@
 
 import * as Immutable from 'vendor/immutable';
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {MultiDict} from '../../util';
 import {parse as parseQuery} from 'cqlParser/parser';
 import {TextTypesStore} from '../textTypes/attrValues';
