+++ public/files/js/stores/query/withinBuilder.ts
@@ -22,7 +22,7 @@
 
 import * as Immutable from 'vendor/immutable';
 import {SimplePageStore} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 
 /**
  *
