+++ public/files/js/stores/wordlist/form.ts
@@ -27,7 +27,7 @@ import * as Immutable from 'vendor/immutable';
 import * as RSVP from 'vendor/rsvp';
 
 import {SimplePageStore, validateGzNumber} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {MultiDict} from '../../util';
 
 /**
