+++ public/files/js/stores/wordlist/main.ts
@@ -27,7 +27,7 @@
 import * as Immutable from 'vendor/immutable';
 import * as RSVP from 'vendor/rsvp';
 import {SimplePageStore, validateGzNumber} from '../base';
-import {PageModel} from '../../pages/document';
+import {PageModel} from '../../app/main';
 import {WordlistFormStore} from './form';
 import {MultiDict} from '../../util';
 import {WordlistSaveStore} from './save';
